wip #2319 Split emitXxx/tryEmitXxx, more consistent use of hooks
tryEmit methods should mostly not use onErrorDropped/onNextDropped or
onDiscard hooks, leaving that to the emit methods (or to the user).

most usage of sinks in tests relies on `emit` methods, which is pretty
consistent with their new semantic of "doing the right thing" out of the
box (since most test usage didn't even check the Emission result).